remove highlight, rename folder with correct spelling, remove wrapper components in storybook
