+++ packages/story-editor/src/components/panels/document/taxonomies/index.js
