+++ packages/story-editor/src/components/panels/document/taxonomies/test/taxonomies.js
